scripts/local_downloader/requirements.txt
flask==3.0.0
flask-cors==4.0.0
yt-dlp==2024.1.25
youtube-search-python==1.6.6
pydub==0.25.1
requests==2.31.0
python-dotenv==1.0.0
spotipy==2.23.0
watchdog==4.0.0
